+++ src/components/uploaders/image-drop-uploader/hooks/use-drag.js
@@ -11,6 +11,8 @@ var _slicedToArray2 = _interopRequireDefault(require("@babel/runtime/helpers/sli
 
 var _react = require("react");
 
+var _pauseEvent = require("../utils/pause-event");
+
 var useDrag = function useDrag(_ref) {
   var startingPosition = _ref.startingPosition,
       imageDimensions = _ref.imageDimensions,
@@ -61,10 +63,11 @@ var useDrag = function useDrag(_ref) {
     }
   }, [startingPosition, imageDimensions]);
 
-  var handleMouseDown = function handleMouseDown(_ref2) {
-    var clientX = _ref2.clientX,
-        clientY = _ref2.clientY,
-        target = _ref2.target;
+  var handleMouseDown = function handleMouseDown(e) {
+    (0, _pauseEvent.pauseEvent)(e);
+    var clientX = e.clientX,
+        clientY = e.clientY,
+        target = e.target;
 
     if (!isDragging) {
       var rect = target.getBoundingClientRect();
@@ -76,10 +79,11 @@ var useDrag = function useDrag(_ref) {
     }
   };
 
-  var handleMouseMove = function handleMouseMove(_ref3) {
-    var clientX = _ref3.clientX,
-        clientY = _ref3.clientY,
-        target = _ref3.target;
+  var handleMouseMove = function handleMouseMove(e) {
+    (0, _pauseEvent.pauseEvent)(e);
+    var clientX = e.clientX,
+        clientY = e.clientY,
+        target = e.target;
 
     if (isDragging) {
       var rect = target.getBoundingClientRect();
@@ -153,9 +157,9 @@ var useDrag = function useDrag(_ref) {
     setLastTranslation(destination);
   };
 
-  var getDestination = function getDestination(_ref4) {
-    var x = _ref4.x,
-        y = _ref4.y;
+  var getDestination = function getDestination(_ref2) {
+    var x = _ref2.x,
+        y = _ref2.y;
     var destinationX = x;
     var destinationY = y;